+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.9] - 11/10/25
+## [0.1.11] - 11/10/25
+
+### Added
+
+- Finished tests for backgammon game
+
+### Changed
+
+- Changed some details in move_checker, as in some cases it tried to move more than 1 checker.
+- Finished tests for backgammon game
+
+## [0.1.10] - 11/10/25
 
 ### Added